src/requirements.txt
FastAPI
gunicorn
uvicorn
python-decouple
rav
fyers_apiv3
firebase_admin
yfinance
pandas
setuptools
pydantic_settings
dhanhq
